+++ src/types.ts
@@ -1,7 +1,9 @@
 export interface ListTreeIndentationGuidesSettings {
   activeListItemThreading: boolean;
   allBranchesOfActiveBulletListThreading: boolean;
+  bulletThreadingFromNonListHead: boolean;
   enableBulletThreading: boolean;
+  enableStaticListTreeIndentationGuides: boolean;
   bulletThreadingInLivePreview: boolean;
   bulletThreadingInReadingMode: boolean;
   bulletThreadingInSourceMode: boolean;
@@ -9,12 +11,15 @@ export interface ListTreeIndentationGuidesSettings {
   renderInLivePreview: boolean;
   renderInReadingMode: boolean;
   renderInSourceMode: boolean;
+  treatBlankLineSeparatedListBlocksAsOne: boolean;
 }
 
 export const DEFAULT_SETTINGS: ListTreeIndentationGuidesSettings = {
   activeListItemThreading: true,
   allBranchesOfActiveBulletListThreading: false,
+  bulletThreadingFromNonListHead: true,
   enableBulletThreading: false,
+  enableStaticListTreeIndentationGuides: true,
   bulletThreadingInLivePreview: true,
   bulletThreadingInReadingMode: true,
   bulletThreadingInSourceMode: true,
@@ -22,4 +27,5 @@ export const DEFAULT_SETTINGS: ListTreeIndentationGuidesSettings = {
   renderInLivePreview: true,
   renderInReadingMode: true,
   renderInSourceMode: true,
+  treatBlankLineSeparatedListBlocksAsOne: false,
 };